Add The Mailing After Account Creation
Added a function that create a random string of 32 characters
that we use as a unique token in a specific User.
Throws an exception after 10 tries to create a token.
Added phpmailer to send mail, linked it with mailhog, for
test purposes.
So there is a need to do a composer install
to be up to date.
Added the token variable to the import.sql and models for User.