Add Mailing To User After Account Creation
Added the function to create a string of X
random characters, the X value has to be pasted
on funciton calling. This function is used
to to create a unique token for a user.
Added phpmailer, to send an email.
Phpmailer is bound to mailhog for now.
Added the token column in the database.
Added a new exception that is thrown if
we can't create a token.